Official support for RFC 8707: https://www.rfc-editor.org/rfc/rfc8707
We accept this at all points of the flow already. What the RFC outlines is, in effect, that the server should implement policies based on the resources server. Since OA4MP is not tied to any resource server, this is normally done on a per client basis as part of the policy for that client.